feat: add emoji categories and guide sections for markdown reference
- Created `emojiCategories.ts` to define various emoji categories with icons and keywords.
- Developed `guideSections.ts` to provide structured markdown guidance, including headings, formatting, lists, and more.
- Introduced `printStyles.ts` for print media styles and page configurations.
- Implemented end-to-end tests in `editor.spec.ts` to ensure core functionalities and UI elements are working as expected.
- Configured Playwright in `playwright.config.ts` for testing environment setup.
