+++ .gitignore
@@ -12,6 +12,8 @@ node_modules
 # Logs
 logs
 *.log
+playwright-report
+test-results
 
 # Misc
 .DS_Store
