+++ app/components/MarketingHero.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="index-hero">
+    <h1>Write Markdown.<br />Render everything.</h1>
+    <p>
+      A browser-based markdown editor with live preview, Mermaid diagrams, LaTeX math, and syntax
+      highlighting. Free, no account required.
+    </p>
+    <div class="index-hero-actions">
+      <NuxtLink to="/editor" class="index-btn-primary">Open Editor</NuxtLink>
+      <NuxtLink to="/guide" class="index-btn-ghost">View Guide</NuxtLink>
+    </div>
+  </div>
+</template>
