+++ app/pages/print-preview.vue
@@ -1,70 +1,20 @@
 <template>
   <div class="print-preview-wrapper">
-    <!-- Toolbar -->
-    <div class="toolbar print:hidden">
-      <h1 class="text-xl font-bold">Print Preview</h1>
+    <PrintPreviewToolbar
+      :page-size="pageSize"
+      :margins="margins"
+      @update:page-size="onPageSizeChange"
+      @update:margins="onMarginsChange"
+      @print="handlePrint"
+      @close="close"
+    />
 
-      <!-- Settings Controls -->
-      <div class="flex items-center gap-6">
-        <!-- Paper Size -->
-        <div class="flex items-center gap-2">
-          <label class="text-sm text-gray-300">Size:</label>
-          <select
-            v-model="pageSize"
-            class="settings-select"
-            @change="reRenderPreview"
-          >
-            <option value="A4">A4</option>
-            <option value="Letter">Letter</option>
-            <option value="Legal">Legal</option>
-            <option value="A3">A3</option>
-          </select>
-        </div>
+    <div ref="previewContainer" class="preview-container print:hidden" />
 
-        <!-- Margins -->
-        <div class="flex items-center gap-2">
-          <label class="text-sm text-gray-300">Margins:</label>
-          <select
-            v-model="margins"
-            class="settings-select"
-            @change="reRenderPreview"
-          >
-            <option value="narrow">Narrow (10mm)</option>
-            <option value="normal">Normal (20mm)</option>
-            <option value="wide">Wide (30mm)</option>
-          </select>
-        </div>
-      </div>
-
-      <!-- Action Buttons -->
-      <div class="flex gap-4">
-        <button
-          class="px-4 py-2 bg-accent-active hover:bg-accent-primary rounded text-white font-medium flex items-center gap-2"
-          @click="handlePrint"
-        >
-          <PhPrinter :size="20" />
-          Print / Save PDF
-        </button>
-        <button
-          class="px-4 py-2 bg-gray-700 hover:bg-gray-600 rounded text-white font-medium"
-          @click="close"
-        >
-          Close
-        </button>
-      </div>
-    </div>
-
-    <!-- PagedJS Container (for screen preview only) -->
-    <div ref="previewContainer" class="preview-container print:hidden">
-      <!-- PagedJS will inject content here -->
-    </div>
-
-    <!-- Source Content (hidden on screen, shown when printing) -->
     <div id="print-content" class="print-content-source">
-      <div class="prose max-w-none" v-html="content"></div>
+      <div class="prose max-w-none" v-html="content" />
     </div>
 
-    <!-- Loading State -->
     <div v-if="loading" class="loading-overlay print:hidden">
       <div class="bg-white p-6 rounded-lg shadow-xl text-gray-900 flex flex-col items-center">
         <div class="animate-spin rounded-full h-10 w-10 border-b-2 border-blue-600 mb-4"></div>
@@ -75,424 +25,29 @@
 </template>
 
 <script setup lang="ts">
-import { ref, onMounted, nextTick } from 'vue'
-import { useRouter } from 'vue-router'
-import { PhPrinter } from '@phosphor-icons/vue'
-import { Previewer } from 'pagedjs'
-import { renderMermaidDiagrams } from '../utils/markdownItMermaid'
-import { sanitizeHtml } from '../utils/sanitizer'
+  import PrintPreviewToolbar from '../components/PrintPreviewToolbar.vue'
+  import { usePrintPreview } from '../composables/usePrintPreview'
+  import type { MarginSize, PageSize } from '../utils/printStyles'
 
-// Use the print layout
-definePageMeta({
-  layout: 'print'
-})
-
-const router = useRouter()
-const content = ref('')
-const loading = ref(true)
-const previewContainer = ref<HTMLElement | null>(null)
-
-// Page settings
-type PageSize = 'A4' | 'Letter' | 'Legal' | 'A3'
-type MarginSize = 'narrow' | 'normal' | 'wide'
-
-const pageSize = ref<PageSize>('A4')
-const margins = ref<MarginSize>('normal')
-
-// Page size and margin values
-const pageSizes: Record<PageSize, string> = {
-  'A4': 'A4',
-  'Letter': 'letter',
-  'Legal': 'legal',
-  'A3': 'A3'
-}
-const marginValues: Record<MarginSize, string> = {
-  'narrow': '10mm',
-  'normal': '20mm',
-  'wide': '30mm'
-}
-
-const PRINT_CODE_LINES_CLASS = 'print-code-lines'
-const PRINT_CODE_LINE_CLASS = 'print-code-line'
-
-// Generate @page CSS for PagedJS (passed as stylesheet to preview())
-const getPageStylesForPagedJS = () => {
-  return `
-    @page {
-      size: ${pageSizes[pageSize.value]};
-      margin: ${marginValues[margins.value]};
-    }
-
-    /* Typography */
-    .prose {
-      color: #374151;
-      max-width: none;
-      font-size: 11pt;
-      line-height: 1.2;
-    }
-
-    .prose h1, .prose h2, .prose h3, .prose h4, .prose strong, .prose b {
-      color: #111827;
-    }
-
-    .prose a {
-      color: #2563eb;
-      text-decoration: underline;
-    }
-
-    .prose blockquote {
-      color: #4b5563;
-      border-left-color: #e5e7eb;
-    }
-
-    .prose code {
-      color: #111827;
-      background-color: #f3f4f6;
-      border: 1px solid #e5e7eb;
-      font-size: 9pt;
-    }
-
-    .prose pre {
-      background-color: #f9fafb;
-      border: 1px solid #e5e7eb;
-      color: #1f2937;
-      box-shadow: none;
-      font-size: 9pt;
-      line-height: 1.2;
-    }
-
-    .prose pre code {
-      background-color: transparent;
-      border: none;
-      color: inherit;
-      font-size: inherit;
-    }
-
-    .code-block-container {
-      background-color: #ffffff;
-      border: 1px solid #e5e7eb;
-      border-left: 1px solid #e5e7eb;
-      border-radius: 4px;
-      padding: 0.75rem 1rem;
-      overflow: visible;
-      break-inside: auto;
-      page-break-inside: auto;
-      box-decoration-break: clone;
-      -webkit-box-decoration-break: clone;
-    }
-
-    .code-block-body {
-      max-height: none !important;
-      overflow: visible !important;
-      background: transparent;
-      display: block;
-    }
-
-    .code-block-container pre {
-      display: block;
-      margin: 0 !important;
-      padding: 0 !important;
-      background-color: #ffffff !important;
-      border: 0 !important;
-      color: #111827 !important;
-      overflow: visible !important;
-      line-height: 1.45 !important;
-      white-space: break-spaces !important;
-      box-decoration-break: clone;
-      -webkit-box-decoration-break: clone;
-      orphans: 3;
-      widows: 3;
-      tab-size: 2;
-    }
-
-    .code-block-container pre code {
-      display: block;
-      margin: 0;
-      border: 0 !important;
-      white-space: break-spaces !important;
-      overflow-wrap: anywhere !important;
-      word-break: break-word !important;
-      tab-size: 2;
-    }
-
-    .prose pre code.${PRINT_CODE_LINES_CLASS},
-    .code-block-container pre code.${PRINT_CODE_LINES_CLASS} {
-      white-space: normal !important;
-      overflow-wrap: normal !important;
-      word-break: normal !important;
-    }
-
-    .prose pre code.${PRINT_CODE_LINES_CLASS} .${PRINT_CODE_LINE_CLASS},
-    .code-block-container pre code.${PRINT_CODE_LINES_CLASS} .${PRINT_CODE_LINE_CLASS} {
-      display: block;
-      white-space: pre-wrap;
-      overflow-wrap: anywhere;
-      word-break: normal;
-      min-height: 1.45em;
-    }
-
-    .prose table {
-      border-color: #e5e7eb;
-    }
-
-    .prose th {
-      background-color: #f9fafb;
-      color: #111827;
-      border-color: #e5e7eb;
-    }
-
-    .prose td {
-      color: #374151;
-      border-color: #e5e7eb;
-    }
-
-    /* KaTeX Math */
-    .katex, .katex * {
-      color: #111827;
-    }
-
-    /* Task Lists */
-    .prose ul li,
-    .prose ol li,
-    .prose li.task-list-item,
-    .prose .task-list-item {
-      color: #374151;
-    }
-
-    .prose input[type="checkbox"] {
-      accent-color: #2563eb;
-    }
-
-    /* Mermaid Diagrams */
-    .mermaid-container {
-      background-color: transparent;
-      border: none;
-      margin: 1em 0;
-      box-shadow: none;
-    }
-
-    .mermaid-viewport {
-      overflow: visible !important;
-      height: auto !important;
-      min-height: auto !important;
-    }
-
-    .mermaid-diagram {
-      transform: none !important;
-    }
-
-    .mermaid {
-      display: block !important;
-    }
-
-    .mermaid svg {
-      max-width: 100% !important;
-      height: auto !important;
-    }
-
-    .mermaid-controls,
-    .code-block-header {
-      display: none;
-    }
-
-    /* Page breaks */
-    h1, h2, h3, h4, h5, h6 {
-      page-break-after: avoid;
-    }
-
-    img, figure, table {
-      page-break-inside: avoid;
-    }
-  `
-}
-
-// Generate @media print CSS for window.print() (injected into document head)
-const getPrintMediaStyles = () => {
-  return `
-    @media print {
-      @page {
-        size: ${pageSizes[pageSize.value]};
-        margin: 0mm;
-      }
-
-      .pagedjs_page {
-        width: 100% !important;
-        height: 100vh !important;
-        page-break-after: always !important;
-        break-after: page !important;
-      }
-
-      .pagedjs_page:last-child {
-        page-break-after: avoid !important;
-        break-after: avoid !important;
-      }
-    }
-  `
-}
-
-// Store current blob URL for cleanup
-let currentBlobUrl: string | null = null
-
-// Create a blob URL for the PagedJS stylesheet
-const createStylesheetBlobUrl = () => {
-  if (currentBlobUrl) {
-    URL.revokeObjectURL(currentBlobUrl)
-  }
-  const cssContent = getPageStylesForPagedJS()
-  const blob = new Blob([cssContent], { type: 'text/css' })
-  currentBlobUrl = URL.createObjectURL(blob)
-  return currentBlobUrl
-}
-
-// Inject print media styles into document head (for window.print())
-const injectPrintMediaStyles = () => {
-  const existingStyle = document.getElementById('dynamic-print-styles')
-  if (existingStyle) existingStyle.remove()
-
-  const styleEl = document.createElement('style')
-  styleEl.id = 'dynamic-print-styles'
-  styleEl.textContent = getPrintMediaStyles()
-  document.head.appendChild(styleEl)
-}
-
-const loadContent = () => {
-  if (import.meta.client) {
-    document.documentElement.classList.remove('dark')
-    const storedContent = localStorage.getItem('mermditor-print-content')
-    if (storedContent) {
-      content.value = sanitizeHtml(storedContent)
-    } else {
-      // Fallback or redirect if no content
-      router.push('/editor')
-    }
-  }
-}
-
-const normalizeCodeBlocksForPrint = (root: Element | Document) => {
-  const codeBlocks = root.querySelectorAll('pre code')
-
-  for (const codeBlock of codeBlocks) {
-    if (codeBlock.classList.contains('language-mermaid')) {
-      continue
-    }
-
-    if (codeBlock.classList.contains(PRINT_CODE_LINES_CLASS)) {
-      continue
-    }
-
-    const rawText = codeBlock.textContent
-    if (!rawText) {
-      continue
-    }
-
-    const lines = rawText.replace(/\r\n?/g, '\n').split('\n')
-    const fragment = document.createDocumentFragment()
-
-    for (const line of lines) {
-      const lineElement = document.createElement('span')
-      lineElement.className = PRINT_CODE_LINE_CLASS
-      lineElement.textContent = line.length > 0 ? line : '\u00a0'
-      fragment.appendChild(lineElement)
-    }
-
-    codeBlock.textContent = ''
-    codeBlock.classList.add(PRINT_CODE_LINES_CLASS)
-    codeBlock.appendChild(fragment)
-  }
-}
-
-const initializePreview = async () => {
-  if (!import.meta.client) return
-
-  // 1. Inject print media styles for window.print()
-  injectPrintMediaStyles()
-
-  // 2. Wait for Vue to render the source content
-  await nextTick()
-
-  // 3. Render diagrams in the source container
-  await renderMermaidDiagrams({
-    theme: 'default',
-    startOnLoad: false,
-    htmlLabels: false,
-    flowchart: { htmlLabels: false },
-    themeVariables: {
-      primaryColor: '#2563eb', // blue-600
-      primaryTextColor: '#111827', // gray-900
-      primaryBorderColor: '#e5e7eb', // gray-200
-      lineColor: '#374151', // gray-700
-      secondaryColor: '#f3f4f6', // gray-100
-      tertiaryColor: '#ffffff', // white
-      background: '#ffffff',
-      mainBkg: '#ffffff',
-      secondBkg: '#f3f4f6',
-      tertiaryBkg: '#e5e7eb',
-    }
+  definePageMeta({
+    layout: 'print',
   })
 
-  normalizeCodeBlocksForPrint(document.querySelector('#print-content') ?? document)
+  const { content, loading, previewContainer, pageSize, margins, reRenderPreview, handlePrint, close } =
+    usePrintPreview()
 
-  // 4. Run Paged.js with our complete stylesheet
-  const paged = new Previewer()
-  const source = document.querySelector('#print-content')
-
-  if (source && previewContainer.value) {
-    // We need to clear previous content if any
-    previewContainer.value.innerHTML = ''
-
-    // Listen for the rendered event to hide loading
-    paged.on('rendered', () => {
-      loading.value = false
-    })
-
-    // Start preview
-    // Pass our complete stylesheet as a blob URL - this avoids PagedJS parsing
-    // all document stylesheets (which can fail on complex Tailwind selectors)
-    const stylesheetUrl = createStylesheetBlobUrl()
-    await paged.preview(source.innerHTML, [stylesheetUrl], previewContainer.value)
+  const onPageSizeChange = async (value: PageSize): Promise<void> => {
+    pageSize.value = value
+    await reRenderPreview()
   }
-}
 
-// Re-render preview when settings change
-const reRenderPreview = async () => {
-  loading.value = true
-  if (previewContainer.value) {
-    previewContainer.value.innerHTML = ''
+  const onMarginsChange = async (value: MarginSize): Promise<void> => {
+    margins.value = value
+    await reRenderPreview()
   }
-  await initializePreview()
-}
-
-const handlePrint = () => {
-  // Set document title for default PDF filename
-  const timestamp = Date.now()
-  const originalTitle = document.title
-  document.title = `document - ${timestamp}`
-
-  window.print()
-
-  // Restore original title after print dialog closes
-  document.title = originalTitle
-}
-
-const close = () => {
-  window.close()
-  if (!window.closed) {
-    // If opened in same tab or window.close() blocked
-    router.back()
-  }
-}
-
-onMounted(async () => {
-  loadContent()
-  // Add a small delay to ensure DOM is ready and styles are loaded
-  setTimeout(() => {
-    initializePreview()
-  }, 500)
-})
 </script>
 
 <style>
-/* Screen styles for the print preview page */
 .print-preview-wrapper {
   min-height: 100vh;
   display: flex;
@@ -507,7 +62,7 @@ onMounted(async () => {
   left: 0;
   right: 0;
   z-index: 50;
-  background-color: rgb(17 24 39); /* gray-900 */
+  background-color: rgb(17 24 39);
   color: white;
   padding: 1rem;
   box-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1);
@@ -517,8 +72,8 @@ onMounted(async () => {
 }
 
 .settings-select {
-  background-color: rgb(55 65 81); /* gray-700 */
-  border: 1px solid rgb(75 85 99); /* gray-600 */
+  background-color: rgb(55 65 81);
+  border: 1px solid rgb(75 85 99);
   color: white;
   padding: 0.375rem 0.75rem;
   border-radius: 0.375rem;
@@ -528,11 +83,11 @@ onMounted(async () => {
 }
 
 .settings-select:hover {
-  background-color: rgb(75 85 99); /* gray-600 */
+  background-color: rgb(75 85 99);
 }
 
 .settings-select:focus {
-  border-color: rgb(59 130 246); /* blue-500 */
+  border-color: rgb(59 130 246);
   box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.3);
 }
 
@@ -554,9 +109,7 @@ onMounted(async () => {
   z-index: 40;
 }
 
-/* Critical print styles - these override everything during print */
 @media print {
-  /* Reset the wrapper */
   .print-preview-wrapper {
     min-height: auto !important;
     background: white !important;
@@ -565,17 +118,14 @@ onMounted(async () => {
     margin: 0 !important;
   }
 
-  /* Hide toolbar */
   .toolbar {
     display: none !important;
   }
 
-  /* Hide loading overlay */
   .loading-overlay {
     display: none !important;
   }
 
-  /* Show the PagedJS preview container during print */
   .preview-container {
     display: block !important;
     margin: 0 !important;
@@ -583,7 +133,6 @@ onMounted(async () => {
     width: 100% !important;
   }
 
-  /* Hide the source content - use PagedJS rendered pages */
   .print-content-source {
     display: none !important;
   }
