+++ app/utils/emojiCategories.ts
@@ -1,0 +1,255 @@
+export interface EmojiCategory {
+  name: string
+  icon: string
+  keywords: string[]
+}
+
+export const emojiCategories: EmojiCategory[] = [
+  {
+    name: 'Faces',
+    icon: '😀',
+    keywords: [
+      'smile',
+      'happy',
+      'joy',
+      'laugh',
+      'grin',
+      'blush',
+      'wink',
+      'cry',
+      'angry',
+      'love',
+      'heart_eyes',
+      'kiss',
+      'tongue',
+      'sleeping',
+      'mask',
+      'sunglasses',
+      'confused',
+      'worried',
+      'relieved',
+      'satisfied',
+      'neutral',
+      'expressionless',
+      'unamused',
+      'sweat',
+      'weary',
+      'pensive',
+      'disappointed',
+      'confounded',
+      'fearful',
+      'persevere',
+      'sob',
+      'astonished',
+      'scream',
+      'tired',
+      'rage',
+      'triumph',
+      'sleepy',
+      'yum',
+      'dizzy',
+      'imp',
+      'smiling_imp',
+      'innocent',
+      'alien',
+    ],
+  },
+  {
+    name: 'Gestures',
+    icon: '👍',
+    keywords: [
+      'thumbsup',
+      'thumbsdown',
+      'ok_hand',
+      'punch',
+      'fist',
+      'wave',
+      'hand',
+      'raised_hand',
+      'open_hands',
+      'point_up',
+      'point_down',
+      'point_left',
+      'point_right',
+      'raised_hands',
+      'pray',
+      'clap',
+      'muscle',
+      'metal',
+      'walking',
+      'runner',
+      'running',
+      'dancer',
+      'dancers',
+    ],
+  },
+  {
+    name: 'Hearts',
+    icon: '❤️',
+    keywords: [
+      'heart',
+      'yellow_heart',
+      'blue_heart',
+      'purple_heart',
+      'green_heart',
+      'broken_heart',
+      'heartbeat',
+      'heartpulse',
+      'two_hearts',
+      'revolving_hearts',
+      'cupid',
+      'sparkling_heart',
+    ],
+  },
+  {
+    name: 'Animals',
+    icon: '🐶',
+    keywords: [
+      'dog',
+      'cat',
+      'mouse',
+      'hamster',
+      'rabbit',
+      'wolf',
+      'frog',
+      'tiger',
+      'koala',
+      'bear',
+      'pig',
+      'cow',
+      'boar',
+      'monkey',
+      'horse',
+      'camel',
+      'sheep',
+      'elephant',
+      'panda',
+      'snake',
+      'bird',
+      'chicken',
+      'penguin',
+      'turtle',
+      'bug',
+      'honeybee',
+      'ant',
+      'beetle',
+      'snail',
+      'octopus',
+      'fish',
+      'whale',
+      'dolphin',
+    ],
+  },
+  {
+    name: 'Food',
+    icon: '🍎',
+    keywords: [
+      'apple',
+      'orange',
+      'banana',
+      'grapes',
+      'strawberry',
+      'pizza',
+      'hamburger',
+      'fries',
+      'coffee',
+      'tea',
+      'beer',
+      'wine',
+      'cake',
+      'cookie',
+      'chocolate',
+      'candy',
+      'bread',
+      'egg',
+      'meat',
+      'rice',
+      'sushi',
+      'ice_cream',
+    ],
+  },
+  {
+    name: 'Activities',
+    icon: '⚽',
+    keywords: [
+      'soccer',
+      'basketball',
+      'football',
+      'tennis',
+      'golf',
+      'baseball',
+      'rugby',
+      'bowling',
+      'ski',
+      'snowboard',
+      'swimming',
+      'surfing',
+      'bike',
+      'car',
+      'airplane',
+      'rocket',
+      'guitar',
+      'music',
+      'art',
+      'microphone',
+      'headphones',
+    ],
+  },
+  {
+    name: 'Travel',
+    icon: '🚗',
+    keywords: [
+      'car',
+      'bus',
+      'train',
+      'airplane',
+      'rocket',
+      'ship',
+      'bike',
+      'taxi',
+      'police_car',
+      'fire_engine',
+      'ambulance',
+      'truck',
+      'hotel',
+      'house',
+      'office',
+      'school',
+      'hospital',
+      'bank',
+      'church',
+    ],
+  },
+  {
+    name: 'Symbols',
+    icon: '⭐',
+    keywords: [
+      'star',
+      'fire',
+      'sparkles',
+      'boom',
+      'star2',
+      'dizzy',
+      'anger',
+      'exclamation',
+      'question',
+      'zzz',
+      'dash',
+      'sweat_drops',
+      'notes',
+      'musical_note',
+      'heavy_check_mark',
+      'x',
+      'heavy_plus_sign',
+      'heavy_minus_sign',
+      'heavy_multiplication_x',
+      'heavy_division_sign',
+      'arrow_up',
+      'arrow_down',
+      'arrow_left',
+      'arrow_right',
+      'warning',
+      'information_source',
+    ],
+  },
+]
